test: Generic dummy dev layer build script
