--- node/packages/aws-lambda-sdk/test/lib/build-dummy-dev-mode-extension.js
@@ -1,15 +1,0 @@
-'use strict';
-
-const path = require('path');
-const fsp = require('fs').promises;
-const mkdir = require('fs2/mkdir');
-const resolveDirZipBuffer = require('../../../../test/utils/resolve-dir-zip-buffer');
-
-const zipFilename = path.resolve(__dirname, '../../dist/dummy-dev-mode.zip');
-const extensionDirname = path.resolve(__dirname, '../fixtures/dev-mode-extension');
-
-module.exports = async () => {
-  await mkdir(path.dirname(zipFilename), { intermediate: true, silent: true });
-  await fsp.writeFile(zipFilename, await resolveDirZipBuffer(extensionDirname));
-  return zipFilename;
-};
