--- node/packages/aws-lambda-sdk/test/scripts/build-dummy-dev-mode-extension.js
@@ -1,8 +1,0 @@
-#!/usr/bin/env node
-
-'use strict';
-
-require('essentials');
-require('log-node')();
-
-require('../lib/build-dummy-dev-mode-extension')();
