+++ node/test/lib/layers/dummy-dev-mode-runtime-agnostic/extensions/dummy-dev-mode.js
@@ -1,0 +1,6 @@
+#!/opt/dummy-dev-mode/node
+
+'use strict';
+
+// eslint-disable-next-line import/no-unresolved
+require('../dummy-dev-mode');
