+++ node/test/lib/layers/dummy-dev-mode/extensions/dummy-dev-mode.js
@@ -1,0 +1,6 @@
+#!/usr/bin/env node
+
+'use strict';
+
+// eslint-disable-next-line import/no-unresolved
+require('../dummy-dev-mode');
